starting to write the hierarchical system build script
