+++ pyext/src/Files.cmake
@@ -1,3 +1,3 @@
-set(pyfiles "restraints/strings.py")
+set(pyfiles "restraints/strings.py;samplers/two_level_mcmc.py")
 set(cppfiles "")
 set(cudafiles "")
